Updated monetary references to use decimal instead o integer

+++ Service/Stocks.Domain/Aggregates/AccountAggregate/Account.cs
@@ -15,7 +15,7 @@ namespace Stocks.Domain.Aggregates.AccountAggregate {
         /// <summary>
         /// Current available cash of the account.
         /// </summary>
-        public int Cash { get; protected set; }
+        public decimal Cash { get; protected set; }
 
         /// <summary>
         /// Current share balances for this account
@@ -35,7 +35,7 @@ namespace Stocks.Domain.Aggregates.AccountAggregate {
         /// </summary>
         /// <param name="initialCash">Initial balance amount for the account.</param>
         /// <exception cref="InvalidInitialBalanceException">Thrown if the initial balance of the account is less than zero.</exception>
-        public Account(int initialCash) : this() {
+        public Account(decimal initialCash) : this() {
             if (initialCash < 0)
                 throw new InvalidInitialBalanceException();
 
@@ -46,7 +46,7 @@ namespace Stocks.Domain.Aggregates.AccountAggregate {
         /// Increase the balance of the account by the specified amount.
         /// </summary>
         /// <param name="amount">Amount to deposit.</param>
-        public void Deposit(int amount) {
+        public void Deposit(decimal amount) {
             if (amount < 0)
                 throw new InvalidAccountOperationException("The amount to deposit cannot be negative.");
             Cash += amount;
@@ -57,7 +57,7 @@ namespace Stocks.Domain.Aggregates.AccountAggregate {
         /// </summary>
         /// <param name="amount">Amount to deduct.</param>
         /// <exception cref="InsufficientBalanceException">Thrown if the current balance is not enough to cover the deduction.</exception>
-        public void Deduct(int amount) {
+        public void Deduct(decimal amount) {
             if (amount < 0)
                 throw new InvalidAccountOperationException("The amount to deduct must be positive to be deducted.");
             if (Cash < amount)
@@ -68,7 +68,7 @@ namespace Stocks.Domain.Aggregates.AccountAggregate {
         /// <summary>
         /// Execute stock order.
         /// </summary>
-        public Transaction PlaceOrder(DateTime timestamp, Operation operation, string issuer, int shares, int sharePrice) {
+        public Transaction PlaceOrder(DateTime timestamp, Operation operation, string issuer, int shares, decimal sharePrice) {
             var transaction = new Transaction(this, timestamp, operation, issuer, shares, sharePrice);
             var balance = _shareBalances.FirstOrDefault(_ => _.Issuer == issuer);
